ALGORITHM Refactor to set algorithm
This is where some fun will be had.

Need to check each letter.

If it's a type, and there's only one, add another different of that
type.
If it's a type, and there's more than one and they're all the same, add
another different of that type.
If it's not a type, there's a slim chance to duplicate one of the
letters from the word.
If not, randomly choose one of the types and then a random of that
type.